agents/MLEvolve/requirements_domain.txt
opencv-python-headless==4.10.0.84
albumentations==2.0.5
albucore==0.0.23
imgaug==0.4.0
kornia==0.8.0
kornia_rs==0.1.8
pillow_heif==0.22.0
imageio==2.37.0
imagecodecs==2024.12.30
tifffile==2025.3.13
tiffslide==2.5.1
pdf2image==1.17.0


ultralytics==8.3.221
ultralytics-thop==2.0.14
yolov5==7.0.14
effdet==0.3.0
efficientnet_pytorch==0.6.3
segmentation_models_pytorch==0.5.0
pretrainedmodels==0.7.4
ensemble-boxes==1.0.9
pybboxes==0.1.6
pycocotools==2.0.8
roboflow==1.1.66
sahi==0.11.23
ttach==0.0.3
fvcore==0.1.5.post20221221
thop==0.1.1.post2209072238
open_clip_torch==3.2.0
x-clip==0.14.4


monai==1.4.0
pydicom==3.0.1
gdcm==1.1
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
simpleitk==2.5.3
nibabel==5.3.2
openslide-bin==4.0.0.11
openslide-python==1.4.2
torchxrayvision==1.3.4
torchio==0.21.0


spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
thinc==8.3.4
blis==1.2.0
cymem==2.0.11
preshed==3.0.9
srsly==2.5.1
wasabi==1.1.3
weasel==0.4.1
nltk==3.9.1
gensim==4.3.3
fasttext==0.9.3
kenlm==0.3.0
pyctcdecode==0.5.0
nlpaug==1.1.0
googletrans==4.0.2
sacremoses==0.1.1
Morfessor==2.0.6
indic-nlp-library==0.92
transliterate==1.10.2


emoji==0.6.0
contractions==0.1.73
textsearch==0.0.24
anyascii==0.3.2
textblob==0.19.0
textstat==0.7.5
pyphen==0.17.2
cmudict==1.0.32
num2words==0.5.10
inflect==7.5.0
fuzzywuzzy==0.18.0
editdistance==0.8.1
pyahocorasick==2.1.0
pyspellchecker==0.8.4
pytesseract==0.3.13
pyocr==0.8.5
python-crfsuite==0.9.11
sklearn-crfsuite==0.5.0
markovify==0.9.4
vaderSentiment==3.3.2
rank-bm25==0.2.2
bibtexparser==1.4.3


librosa==0.9.2
audioread==3.0.1
soundfile==0.13.1
resampy==0.4.3
soxr==0.5.0.post1
nnAudio==0.3.3
noisereduce==3.0.3
torchlibrosa==0.1.0
torchvggish==0.2
panns-inference==0.1.1
python_speech_features==0.6


sktime==0.40.1
gluonts==0.16.2
statsforecast==2.0.1
mlforecast==0.14.0
chronos-forecasting==1.5.2
coreforecast==0.0.16
utilsforecast==0.2.11
window_ops==0.0.15
hmmlearn==0.3.3
pykalman==0.11.0
holidays==0.86
dtaidistance==2.3.13
fastdtw==0.3.4
bayespy==0.5.1


dgl==2.1.0
torch-geometric==2.6.1
torchsort
torch-dct==0.1.6
igraph==0.11.8

# Needs to be downloaded in https://pytorch-geometric.com/whl/torch-2.7.0%2Bcu128.html
# torch_cluster==1.6.3
# torch_scatter==2.1.2
# torch_sparse==0.6.18
# torch_spline_conv==1.2.2

faiss-cpu==1.11.0
annoy==1.17.3
nmslib==2.1.2
implicit==0.7.2
simsimd==6.2.1
pynndescent==0.5.13


pytorch-crf==0.7.2
TorchCRF==1.1.0
pytorch-metric-learning==2.8.1
pytorch-msssim==1.0.0
pytorch_ssim==0.1
piq==0.8.0
piqa==1.3.2
lion-pytorch==0.2.3
ranger21==0.1.0
performer-pytorch==1.1.4    
local-attention==1.11.2
axial_positional_embedding==0.3.12
hyper-connections==0.2.1
torch_discounted_cumsum==1.1.0
opt-einsum-fx==0.1.4


rasterio==1.4.3
affine==2.4.0
pyproj==3.7.1
shapely==2.0.7
cligj==0.7.2
geographiclib==2.0
geopy==2.4.1
haversine==2.9.0
pymap3d==3.1.0
pyquaternion==0.9.9
numpy-quaternion==2024.0.13


rdkit==2025.3.5
biopython==1.85
dscribe==2.1.1
e3nn==0.5.0


open3d==0.19.0


opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
modelscope==1.25.0
model-index==0.1.11


PyWavelets==1.8.0
lmdb==1.7.5
jpegio==0.2.8
iopath==0.1.9
iterative-stratification==0.1.9
seqeval==1.2.2
adagio==0.2.6
addict==2.4.0
antlr4-python3-runtime==4.9.3
fastai==2.8.0
fastcore==1.8.0
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.1
fugue==0.9.1
triad==0.9.8
muq==0.1.0
marisa-trie==1.2.1
pygtrie==2.5.0
pydeck==0.9.1
pathos==0.3.4
pox==0.3.6
ppft==1.7.7
py4j==0.10.9.9
umap==0.1.1
umap-learn==0.5.9.post2
yacs==0.1.8
alembic==1.15.1
pymongo==4.11.2
